ndr_sensor/requirements.txt
fastapi
uvicorn
requests
python-dotenv
apscheduler
netifaces
scapy
